aded local storage

+++ QA-ui/QA-ui/src/app/questions/delete-questions/delete-questions.component.ts
@@ -9,33 +9,39 @@ import Swal from 'sweetalert2';
 })
 export class DeleteQuestionsComponent implements OnInit {
 
-  constructor(private questionService:QuestionService){}
+  constructor(private questionService: QuestionService) { }
 
   ngOnInit(): void {
-   let id = window.location.pathname.split("/").pop();
-       Swal.fire({
-      title: 'Are you sure want to delete the question?',
-      text: "You won't be able to revert this!",
-      icon: 'warning',
-      showCancelButton: true,
-      confirmButtonColor: '#3085d6',
-      cancelButtonColor: '#d33',
-      confirmButtonText: 'Yes, delete it!'
-    }).then((result) => {
-      if (result.isConfirmed) {
-        Swal.fire(
-          'Deleted!',
-          'Question has been deleted.',
-          'success'
-        ).then(() =>{
-          this.questionService.deleteQuestion(id).subscribe((data: any) => {
 
-           window.location.href = "http://localhost:4200/questions/userquestion";
+    var logingStatus = localStorage.getItem('token')
+    if (logingStatus) {
+      let id = window.location.pathname.split("/").pop();
+      Swal.fire({
+        title: 'Are you sure want to delete the question?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire(
+            'Deleted!',
+            'Question has been deleted.',
+            'success'
+          ).then(() => {
+            this.questionService.deleteQuestion(id).subscribe((data: any) => {
+
+              window.location.href = "http://localhost:4200/questions/userquestion";
+            })
           })
-        })
-      }else{
-        window.location.href = "http://localhost:4200/questions/userquestion";
-      }
-    })
+        } else {
+          window.location.href = "http://localhost:4200/questions/userquestion";
+        }
+      })
+    } else {
+      window.location.href = "http://localhost:4200/login";
+    }
   }
 }
